Enable DHCP with MIP tests. Separate sources for high and low level stack tests. Fixed assertion.
